Add Cloverleaf
**https://cloverleaf.app**

**An open source app to replace your password manager without storing your passwords anywhere.**

Cloverleaf is a web app that I've been working on for 4 years now with the hope of making using secure passwords easy and painless. All you need to do is enter the name of the app you want to make a password for and your super secret master password. This'll then give you a password that fits all that site's requirements without storing it anywhere.